Ensure player inventory reflects cards + items that are removed from their deck
Prior to this commit; if a key is consumed in the dungeon but was
somehow still in the player's inventory in the lobby, then the key will
still remain in the player's lobby inventory when they reconnect. This
was because Citadel did not remove items/cards from the player's
inventory if they were removed from the deck.

Relates to #45